image_api/requirements.txt
django==3.2.6
djangorestframework==3.12.4
Pillow==8.3.2
